requirements.txt
bcolz==1.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
easydict==1.9
et-xmlfile==1.0.1
graphviz==0.8.4
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy 
opencv-python==4.4.0.46
openpyxl==3.0.7
pandas==1.1.5
Pillow 
protobuf==3.15.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.3.3
sklearn==0.0
tensorboardX==2.1
threadpoolctl==2.1.0
torch==1.8.0
torchvision==0.9.0
tqdm==4.23.4
urllib3==1.26.3
